requirements.txt
# ------------------------------------------------------------------------------
#  Numenta Platform for Intelligent Computing (NuPIC)
#  Copyright (C) 2019, Numenta, Inc.  Unless you have an agreement
#  with Numenta, Inc., for a separate license for this software code, the
#  following terms and conditions apply:
#
#  This program is free software: you can redistribute it and/or modify
#  it under the terms of the GNU Affero Public License version 3 as
#  published by the Free Software Foundation.
#
#  This program is distributed in the hope that it will be useful,
#  but WITHOUT ANY WARRANTY; without even the implied warranty of
#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
#  See the GNU Affero Public License for more details.
#
#  You should have received a copy of the GNU Affero Public License
#  along with this program.  If not, see http://www.gnu.org/licenses.
#
#  http://numenta.org/licenses/
#
# ------------------------------------------------------------------------------
# Install nupic.torch and nupic.tensorflow directly from github master branch
nupic.torch @ git+https://github.com/numenta/nupic.torch.git
nupic.tensorflow @ git+https://github.com/numenta/nupic.tensorflow.git

awscli
boto3
elasticsearch
jupyter
librosa==0.7.1
matplotlib
numpy
pandas
python-dateutil==2.8.0
fastai
pillow==6.0.0
python-dotenv
ray[dashboard,debug]==0.7.5
requests
scikit-image
seaborn>=0.9.0
tabulate
torchvision
tqdm